shuffle filestructure for more consistency
