+++ amivapi/events/emails.py
@@ -2,21 +2,15 @@
 #
 # license: AGPLv3, see LICENSE for details. In addition we strongly encourage
 #          you to buy us beer if we meet and you like the software.
-"""Email confirmation logic.
+"""Email formatting.
 
-Needed when external users want to sign up for public events.
+Needed when users are notified about their event signups.
 """
-from bson import ObjectId
-from eve.methods.delete import deleteitem_internal
-from eve.methods.patch import patch_internal
-from flask import Blueprint, current_app, redirect, url_for
-from itsdangerous import BadSignature, URLSafeSerializer
+from flask import current_app, url_for
+from itsdangerous import URLSafeSerializer
 
-from amivapi.events.queue import update_waiting_list
 from amivapi.events.utils import get_token_secret
 from amivapi.utils import mail
-
-email_blueprint = Blueprint('emails', __name__)
 
 
 def notify_signup_accepted(event, signup):
@@ -42,20 +36,13 @@ def notify_signup_accepted(event, signup):
     deletion_link = url_for('emails.on_delete_signup', token=token,
                             _external=True)
 
-    title = event.get('title_en') or event.get('title_de')
-    mail(current_app.config['API_MAIL'], email,
-         '[AMIV] Eventsignup accepted',
-         'Hello %s!\n'
-         '\n'
-         'We are happy to inform you that your signup for %s was accepted and '
-         'you can come to the event! If you do not have time to attend the '
-         'event please click this link to free your spot for someone else:\n'
-         '\n%s\n'
-         'You cannot sign out of this event after %s.\n\n'
-         'Best Regards,\n'
-         'The AMIV event bot'
-         % (name, title, deletion_link,
-            event['time_register_end'].strftime('%M.%H %d.%m.%Y')))
+    mail(current_app.config['API_MAIL'], [email],
+         'AMIV Eventsignup accepted',
+         current_app.config['ACCEPT_EMAIL_TEXT'].format(
+             name=name,
+             title=event.get('title_en') or event.get('title_de'),
+             link=deletion_link,
+             deadline=event['time_register_end'].strftime('%H.%M %d.%m.%Y')))
 
 
 def send_confirmmail_to_unregistered_users(items):
@@ -70,10 +57,7 @@ def send_confirmmail_to_unregistered_users(items):
                 'events', None,
                 **{current_app.config['ID_FIELD']: item['event']})
 
-            if event.get('title_en') is not None:
-                title = event['title_en']
-            else:
-                title = event['title_de']
+            title = event.get('title_en') or event.get('title_de')
 
             s = URLSafeSerializer(get_token_secret())
             token = s.dumps(str(item['_id']))
@@ -82,74 +66,11 @@ def send_confirmmail_to_unregistered_users(items):
                 current_app.logger.warning("SERVER_NAME is not set. E-Mail "
                                            "links will not work!")
 
-            fields = {
-                'link': url_for('emails.on_confirm_email', token=token,
-                                _external=True),
-                'title': title
-            }
-            email_content = current_app.config['CONFIRM_EMAIL_TEXT'] % fields
+            confirm_link = url_for('emails.on_confirm_email', token=token,
+                                   _external=True)
 
-            mail(current_app.config['API_MAIL'],  # from
-                 [item['email']],  # receivers list
+            mail(current_app.config['API_MAIL'], [item['email']],
                  'Registration for AMIV event %s' % title,
-                 email_content)
-
-
-def add_confirmed_before_insert(items):
-    """Add the confirmed field to a event signup before it is inserted to the
-    database. We accept all registered users instantly, others need to click the
-    confirmation link first"""
-    for item in items:
-        if item.get('user', None) is None:
-            item['confirmed'] = False
-        else:
-            item['confirmed'] = True
-
-
-@email_blueprint.route('/confirm_email/<token>')
-def on_confirm_email(token):
-    """Email confirmation endpoint.
-
-    We try to confirm the specified signup and redirect to a webpage.
-    """
-    try:
-        s = URLSafeSerializer(get_token_secret())
-        signup_id = ObjectId(s.loads(token))
-    except BadSignature:
-        return "Unknown token"
-
-    patch_internal('eventsignups', {'confirmed': True},
-                   skip_validation=True, concurrency_check=False,
-                   **{current_app.config['ID_FIELD']: signup_id})
-
-    # Now the user may be able to get accepted, so update the events waiting
-    # list
-    lookup = {current_app.config['ID_FIELD']: signup_id}
-    signup = current_app.data.find_one('eventsignups', None, **lookup)
-
-    update_waiting_list(signup['event'])
-
-    redirect_url = current_app.config.get('EMAIL_CONFIRMED_REDIRECT')
-    if redirect_url:
-        return redirect(redirect_url)
-    else:
-        return current_app.config['CONFIRM_TEXT']
-
-
-@email_blueprint.route('/delete_signup/<token>')
-def on_delete_signup(token):
-    """Endpoint to delete signups via email"""
-    try:
-        s = URLSafeSerializer(get_token_secret())
-        signup_id = ObjectId(s.loads(token))
-    except BadSignature:
-        return "Unknown token"
-
-    deleteitem_internal('eventsignups', concurrency_check=False,
-                        **{current_app.config['ID_FIELD']: signup_id})
-
-    redirect_url = current_app.config.get('SIGNUP_DELETED_REDIRECT')
-    if redirect_url:
-        return redirect(redirect_url)
-    else:
-        return current_app.config['SIGNUP_DELETED_TEXT']
+                 current_app.config['CONFIRM_EMAIL_TEXT'].format(
+                     title=title,
+                     link=confirm_link))
